modified the seed and added the script to populate the database.
