+++ README.md
@@ -1,6 +1,6 @@
 # Reviews_staybnb
 Working on SDC Project
-# To populate mongo database with 10M records: 1. npm run seed-mongo 2. npm run populate
+# To populate mongo database with 10M records: npm run seed-mongo.
 # To populate sequelize database with 10M records: npm run seed-sequelize
